requirements.txt
python-chess>=1.999
torch>=2.0
numpy>=1.23
tensorboard
